Significantly optimize search
At app start we put all pokemon into one of 26 lists - a list of pokemon starting with that letter. That allows up to split the search from being among 900-ish pokemon to being among 80-ish, as the user types the first letter. Additionally, matching is not case-insensitive and MUCH faster.

Co-authored-by: Dmitry Grinberg <dmitrygr@gmail.com>